requirements.txt
pyTelegramBotAPI==3.7.1
psycopg2-binary==2.8.5
redis==3.5.0